20260901v35：HUD状态栏+世界时间管理器+网格投影格子化+安全屋/制作体系完善——①HUD状态栏：纯文字6项（生命/精力/饱食/水分/时间/地点），数字去标签+颜色区分（生命红/精力金/饱食绿/水分蓝），事件驱动+轮询兜底双保险实时刷新，饱食/水分显示一位小数；②世界时间管理器（新 Services）：统一推进/结算/跳时/时钟同步/状态修改/事件发布——挂机整点结算（饱食2/水分2）+制作动画跳时+睡觉跳时全走管理器，删除旧游戏时钟（内嵌驱动由GameBootstrap自动挂载，全局判重防双时钟），结算余数按比例折算；③饱食/水分 int→float（精确0.1），生存状态变化事件 float 化，存档 v3→v4 旧档重置；④网格落点投影改格子投影（逐格定位含块偏移，口袋缝隙天然留空，替代单矩形——强制禁全网格红框保留）；⑤制作系统：配方加消耗倍率（0.5倍数，33条配方全部配置），消耗=基础2×倍率×时数（float），制作动画（进度条+时间流逝+HUD时钟同步，时长=时间分/10现实秒，制作中锁定/取消回滚材料精力时间），0级破损家具可用但耗时×2（床/灶台效果减半，储物箱不缩水仓库行防物品丢失）；⑥修复：容器允许类型取消限制（弹挂/腰封/背包通吃）、网格右边界竖线/底边横线块偏移对齐、世界时间不推进（时钟未挂载）修复

+++ Assets/Scripts/Events/事件定义.cs
@@ -32,7 +32,7 @@
         public 金币变化事件(int 当前, int 变化量) { this.当前 = 当前; this.变化量 = 变化量; }
     }
 
-    // 时间变化：游戏内分钟数推进（游戏时钟 每分钟发布一次，UI 据此刷新）
+    // 时间变化：游戏内分钟数推进（世界时间管理器 整点结算/跳时 后发布，UI 据此刷新）
     public readonly struct 时间变化事件
     {
         public readonly float 游戏分钟数;
@@ -57,11 +57,11 @@
             : this(体力, 力量, 智力, 敏捷, 5, 自由属性点) { }
     }
 
-    // 生存状态变化：饱食度/水分度（HUD 与状态栏刷新）
+    // 生存状态变化：饱食度/水分度（HUD 与状态栏刷新；float 精确 0.1）
     public readonly struct 生存状态变化事件
     {
-        public readonly 生存状态类型 类型; public readonly int 当前; public readonly int 变化量;
-        public 生存状态变化事件(生存状态类型 类型, int 当前, int 变化量) { this.类型 = 类型; this.当前 = 当前; this.变化量 = 变化量; }
+        public readonly 生存状态类型 类型; public readonly float 当前; public readonly float 变化量;
+        public 生存状态变化事件(生存状态类型 类型, float 当前, float 变化量) { this.类型 = 类型; this.当前 = 当前; this.变化量 = 变化量; }
     }
 
     // 伤病变化：疲劳/中毒/感冒/流血/骨折/发烧（严重度 0~100，0=无）
